Moved bunch class into utils misc.
