+++ phy/cluster/manual/_update_info.py
@@ -6,9 +6,7 @@
 # Imports
 #------------------------------------------------------------------------------
 
-import numpy as np
-
-from ...utils._bunch import Bunch
+from ...utils._misc import Bunch
 
 
 #------------------------------------------------------------------------------
