+++ phy/utils/tests/test_misc.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
 
-"""Tests of Bunch."""
+"""Tests of misc utility functions."""
 
 #------------------------------------------------------------------------------
 # Imports
 #------------------------------------------------------------------------------
 
-from .._bunch import Bunch
+from .._misc import Bunch
 
 
 #------------------------------------------------------------------------------
-# Test Bunch
+# Tests
 #------------------------------------------------------------------------------
 
 def test_bunch():
